Updated qspi receiver state machine
It works! Also made some modifications to the test bench to help with development. It wasn't enough to discover the problems I had with the state machine because all components in the test bench are still synchronized to the same clock. However, that isn't the case when receiving QSPI from an outside microprocessor.